Tuning bundle script

+++ make_upload.sh
@@ -1,20 +1,35 @@
 #!/usr/bin/env bash
 
+set -e
+
 # Working directories
 root=$(pwd)
 dist="${root}/dist"
+bundle="${dist}/packaged_$1.zip"
 tmp="${root}/tmp"
 build="${tmp}/$1"
 
+cleanup() {
+  if [ -e "${build}" ]; then
+    rm -rf "${build}"
+  fi
+}
+
+trap cleanup EXIT
+
 # Workspace preparations
-rm -rf "${build}"
+cleanup
 mkdir -p "${dist}"
-rm "${dist}/packaged_$1.zip"
+mkdir -p "${tmp}"
+if [ -e "${bundle}" ]; then
+  rm "${bundle}"
+fi
 
 # Packaging
-(
-  cp -R "$1" "${build}"
-  cd "${build}"
-  cp "${root}/Cargo.toml" .
-  zip -r "${dist}/packaged_$1.zip" *
-)
+echo "working from temp directory: ${build}"
+cp -R "$1" "${build}"
+cd "${build}"
+echo "creating bundle ${bundle} with directory contents"
+zip -r "${bundle}" *
+cd "${root}"
+echo "done"
